Add enable unstable feature libraries
