+++ examples/gui.ts
@@ -1,0 +1,14 @@
+import { Webview } from "https://deno.land/x/webview/mod.ts";
+
+const html = `
+  <html>
+  <body>
+    <h1>Hello from Luwak v${Luwak.version}</h1>
+  </body>
+  </html>
+`;
+
+const webview = new Webview();
+
+webview.navigate(`data:text/html,${encodeURIComponent(html)}`);
+webview.run();
